pyproject.toml
[project]
name = "deepfake-detection-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.1",
    "ipykernel>=7.1.0",
    "kagglehub>=0.4.2",
    "pillow>=12.1.0",
    "streamlit>=1.54.0",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchmetrics>=1.8.2",
    "torchsummary>=1.5.1",
    "torchvision>=0.25.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
